+++ libs/cardano-ledger-pretty/src/Cardano/Ledger/Pretty/Conway.hs
@@ -217,7 +217,7 @@ instance
     let RatifyState {..} = rs
      in ppRecord
           "RatifyState"
-          [ ("EnactState", prettyA rsES)
+          [ ("EnactState", prettyA rsEnactState)
           , ("Future", prettyA rsFuture)
           ]